# contracts/http/openapi.yaml
openapi: 3.1.0
info:
  title: VLXD Management API
  version: 0.1.0
  description: OpenAPI 3.1 contract specification for vlxd backend service.

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Returns backend service health status, version, and server timestamp.
      operationId: getHealth
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /health/ready:
    get:
      summary: Readiness check endpoint
      description: Checks backend service readiness including database pool connectivity.
      operationId: getReadiness
      responses:
        "200":
          description: Service is ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"
        "503":
          description: Service is unready or database is unreachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"

  /api/v1/auth/login:
    post:
      summary: User login and session creation
      description: Authenticates user credentials (email/password), initializes an opaque session, and returns session details with an HttpOnly cookie.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              description: Opaque session cookie (vlxd_session)
              schema:
                type: string
                example: vlxd_session=3a7b...; Path=/; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/auth/logout:
    post:
      summary: User logout and session revocation
      description: Revokes the current session and clears the session cookie.
      operationId: logout
      security:
        - CookieAuth: []
        - BearerAuth: []
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/auth/me:
    get:
      summary: Get current authenticated session profile
      description: Returns information about the currently logged-in user, active tenant, and assigned titles.
      operationId: getAuthMe
      security:
        - CookieAuth: []
        - BearerAuth: []
      responses:
        "200":
          description: Current user session profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMeResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/tenant-users:
    post:
      summary: Invite a user to the current tenant
      operationId: inviteTenantUser
      security:
        - CookieAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteTenantUserRequest"
      responses:
        "201":
          description: Tenant user invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUserResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "409":
          $ref: "#/components/responses/ConflictError"

  /api/v1/tenant-users/{tenantUserId}/status:
    patch:
      summary: Change a tenant user's membership status
      operationId: updateTenantUserStatus
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: tenantUserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantUserStatusRequest"
      responses:
        "200":
          description: Tenant user status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUserResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  /api/v1/tenant-users/{tenantUserId}/titles:
    put:
      summary: Replace a tenant user's assigned titles
      operationId: replaceTenantUserTitles
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: tenantUserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceTenantUserTitlesRequest"
      responses:
        "200":
          description: Tenant user titles replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUserResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: vlxd_session
      description: Opaque session token stored in HttpOnly cookie
    BearerAuth:
      type: http
      scheme: bearer
      description: Opaque session token provided in Authorization header

  headers:
    X-Request-Id:
      description: Unique request trace identifier
      schema:
        type: string
        example: req-123e4567-e89b-12d3-a456-426614174000

  parameters:
    PageQueryParam:
      name: page
      in: query
      description: Page number (1-indexed)
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitQueryParam:
      name: limit
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByQueryParam:
      name: sortBy
      in: query
      description: Field name to sort by
      required: false
      schema:
        type: string
    SortOrderQueryParam:
      name: sortOrder
      in: query
      description: Sort direction
      required: false
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc
    RequestIdHeader:
      name: x-request-id
      in: header
      description: Request trace identifier
      required: false
      schema:
        type: string

  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - version
        - timestamp
      properties:
        status:
          type: string
          enum:
            - ok
          example: ok
        version:
          type: string
          example: 0.1.0
        timestamp:
          $ref: "#/components/schemas/DateTime"

    ReadinessResponse:
      type: object
      required:
        - status
        - database
        - timestamp
      properties:
        status:
          type: string
          enum:
            - ready
            - unready
          example: ready
        database:
          type: string
          enum:
            - connected
            - disconnected
            - disabled
          example: connected
        timestamp:
          $ref: "#/components/schemas/DateTime"

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: admin@vlxd.vn
        password:
          type: string
          minLength: 6
          example: Admin@123456
        tenantCode:
          type: string
          description: Optional tenant code if the user belongs to multiple tenants or specifies tenant explicitly
          example: VLXD-DEFAULT

    AuthUser:
      type: object
      required:
        - id
        - email
        - fullName
        - status
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: admin@vlxd.vn
        phone:
          type: string
          nullable: true
          example: "+84901234567"
        fullName:
          type: string
          example: Quản Trị Viên
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
            - BLOCKED
            - ARCHIVED
          example: ACTIVE

    AuthTenant:
      type: object
      required:
        - id
        - code
        - name
        - status
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440001
        code:
          type: string
          example: VLXD-DEFAULT
        name:
          type: string
          example: Công ty VLXD Mẫu
        status:
          type: string
          enum:
            - ACTIVE
            - SUSPENDED
            - ARCHIVED
          example: ACTIVE

    AuthSession:
      type: object
      required:
        - id
        - expiresAt
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440002
        expiresAt:
          $ref: "#/components/schemas/DateTime"
        createdAt:
          $ref: "#/components/schemas/DateTime"

    AuthTitle:
      type: object
      required:
        - id
        - code
        - name
        - roleGroup
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440003
        code:
          type: string
          example: GDKD
        name:
          type: string
          example: Giám đốc
        roleGroup:
          type: object
          required:
            - id
            - code
            - name
          properties:
            id:
              type: string
              format: uuid
              example: 550e8400-e29b-41d4-a716-446655440004
            code:
              type: string
              example: SUPER_ADMIN
            name:
              type: string
              example: Super admin

    LoginResponse:
      type: object
      required:
        - user
        - tenant
        - session
        - token
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        tenant:
          $ref: "#/components/schemas/AuthTenant"
        session:
          $ref: "#/components/schemas/AuthSession"
        token:
          type: string
          description: Opaque session token string (also set in HttpOnly cookie)
          example: 3a7b5e9f1c...

    LogoutResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Logout successful

    AuthMeResponse:
      type: object
      required:
        - user
        - tenant
        - session
        - isOwner
        - titles
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        tenant:
          $ref: "#/components/schemas/AuthTenant"
        session:
          $ref: "#/components/schemas/AuthSession"
        isOwner:
          type: boolean
          example: true
        titles:
          type: array
          items:
            $ref: "#/components/schemas/AuthTitle"

    InviteTenantUserRequest:
      type: object
      required:
        - email
        - titleIds
      properties:
        email:
          type: string
          format: email
        titleIds:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid

    UpdateTenantUserStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - ACTIVE
            - SUSPENDED
            - REVOKED

    ReplaceTenantUserTitlesRequest:
      type: object
      required:
        - titleIds
      properties:
        titleIds:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid

    TenantUserResponse:
      type: object
      required:
        - id
        - tenantId
        - userId
        - email
        - fullName
        - status
        - titleIds
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
        status:
          type: string
          enum:
            - ACTIVE
            - SUSPENDED
            - REVOKED
        titleIds:
          type: array
          items:
            type: string
            format: uuid

    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - PLAN_LIMIT_REACHED
        - TENANT_NOT_FOUND
        - INVALID_CREDENTIALS
        - USER_SUSPENDED
        - TENANT_SUSPENDED
        - SESSION_EXPIRED
        - SESSION_REVOKED
        - INTERNAL_SERVER_ERROR
        - BAD_REQUEST
      example: VALIDATION_ERROR

    ErrorDetails:
      type: object
      additionalProperties: true
      description: Structured contextual error details or field-level validation errors

    ErrorObject:
      type: object
      required:
        - code
        - message
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          example: Invalid input parameters
        details:
          $ref: "#/components/schemas/ErrorDetails"
        requestId:
          $ref: "#/components/schemas/RequestId"

    ErrorEnvelope:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum:
            - false
          example: false
        error:
          $ref: "#/components/schemas/ErrorObject"

    PaginationQuery:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          default: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        sortBy:
          type: string
        sortOrder:
          type: string
          enum:
            - asc
            - desc
          default: desc

    PaginationMeta:
      type: object
      required:
        - page
        - limit
        - totalItems
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          example: 20
        totalItems:
          type: integer
          minimum: 0
          example: 100
        totalPages:
          type: integer
          minimum: 0
          example: 5

    Money:
      type: integer
      minimum: 0
      description: Monetary amount represented as an integer in VND (Vietnamese Dong)
      example: 250000

    DateTime:
      type: string
      format: date-time
      description: ISO 8601 UTC timestamp format
      example: 2026-08-23T03:00:00.000Z

    RequestId:
      type: string
      description: Unique request trace identifier
      example: req-550e8400-e29b-41d4-a716-446655440000

    OptimisticVersion:
      type: integer
      minimum: 1
      description: Monotonically increasing version counter for optimistic concurrency control (OCC)
      example: 1

  responses:
    BadRequestError:
      description: Bad Request / Validation Failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnauthorizedError:
      description: Authentication credentials missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ForbiddenError:
      description: Insufficient permissions or role capabilities
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFoundError:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ConflictError:
      description: Resource conflict or optimistic locking conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalServerError:
      description: Unexpected internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"